Always allocate exc_stack with GC
This simplifies the code and assumptions surrounding exception stacks a
lot.  It does mean that jl_throw might itself fail with a memory
exception, though it's not clear what alternative we have when the
runtime is really that low on memory. (Though we could reserve it
at an earlier point so that we're more likely to at least report the
root cause correctly.)

TODO: Correct GC buffer management